refactor(e2e): share anthropic cache-control shapes in endpoints_client
